requirements.txt
openai>=1.0.0
spotipy>=2.23.0
pydub>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
